organise loaders for webpack
