+++ package.json
@@ -6,7 +6,8 @@
   "scripts": {
     "build": "babel toolkit --out-dir dist",
     "watch": "babel toolkit --watch --out-dir dist",
-    "prepublish": "npm run build"
+    "prepublish": "npm run build",
+    "test": "echo \"Error: no test specified\" && exit 1"
   },
   "bin": {
     "dev-toolkit": "bin/dev-toolkit.js"
@@ -39,7 +40,7 @@
     "universal-dev-toolkit",
     "dev-toolkit"
   ],
-  "author": "Jorge Antunes",
+  "author": "Jorge Antunes @stoikerty",
   "license": "ISC",
   "bugs": {
     "url": "https://github.com/stoikerty/dev-toolkit/issues"
@@ -79,6 +80,7 @@
     "file-exists": "^1.0.0",
     "file-loader": "^0.9.0",
     "files-require-hook": "^1.0.0",
+    "find-port": "^2.0.1",
     "handlebars": "^4.0.5",
     "handlebars-loader": "^1.3.0",
     "html-webpack-plugin": "^2.21.0",
