+++ toolkit/webpack/config/extensionHooks/index.js
@@ -1,0 +1,2 @@
+export cssHook from './cssHook';
+export filesHook from './filesHook';
