+++ toolkit/webpack/config/loaders/index.js
@@ -1,0 +1,3 @@
+export const eslint from './eslint';
+export const postcss from './postcss';
+export const sass from './sass';
