+++ toolkit/webpack/config/loaders/sass.js
@@ -1,0 +1,7 @@
+import {
+  PATHS,
+} from '../../../_userSettings';
+
+export default {
+  includePaths: PATHS.scssIncludePaths,
+};
